Need to investigate ``` fencing...
For some reason, the ``` for code blocks is bad when rendered through Maruku... or whatever Markdown renderer we're using/Github is using. It would appear to be the "right way" to do a code block, however.

I've switched it to a <pre> ... </pre> block for now, so that Chad and Renaldo's post goes through...