Upding the navbar and resources.
I tweaked the navbar, so that there is an actual "Docs" link. It points to a page that links to many things. It used to point to a page that talked about the Plumbing book and the mailing lists. Now, that resources page points to many things.

The page that used to be about "support" now just points to the mailing list page. This is probably a good idea, since that is how we provide support.

Feel free to edit/suggest changes. I'll migrate some of the resources that are hosted off of c.cc into the repository or onto the server so that they're consolidated in the future. For now, though, this is probably useful to a few people who are exploring things for the first time.